+++ addon/controllers/edit-form.js
@@ -8,9 +8,10 @@ import ErrorableControllerMixin from '../mixins/errorable-controller';
 import FlexberryFileControllerMixin from '../mixins/flexberry-file-controller';
 import needSaveCurrentAgregator from '../utils/need-save-current-agregator';
 import getCurrentAgregator from '../utils/get-current-agregator';
-import PaginatedControllerMixin from 'ember-flexberry/mixins/paginated-controller';
-import ReloadListMixin from 'ember-flexberry/mixins/reload-list-mixin';
+import PaginatedControllerMixin from '../mixins/paginated-controller';
+import ReloadListMixin from '../mixins/reload-list-mixin';
 import SortableControllerMixin from '../mixins/sortable-controller';
+import LimitedControllerMixin from '../mixins/limited-controller';
 import FolvOnEditControllerMixin from '../mixins/flexberry-objectlistview-on-edit-form-controller';
 
 const { getOwner } = Ember;
@@ -52,6 +53,7 @@ FlexberryFileControllerMixin,
 PaginatedControllerMixin,
 ReloadListMixin,
 SortableControllerMixin,
+LimitedControllerMixin,
 FolvOnEditControllerMixin, {
   /**
     Flag to enable return to agregator's path if possible.