+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* `flexberry-objectlistview`:
+    * Now `flexberry-objectlistview` can be placed on `edit-form`.
 
 ## [0.8.0-beta.14] - 2016-12-07
 ### Added